requirements.txt
# TODO: migrate to pip-compile --generate-hashes
# Core Framework
py4web==1.20241127.1
pydal==20260118.1

# Database ORM & Migrations
SQLAlchemy==2.0.36
alembic==1.14.1

# Database Drivers (PostgreSQL, MySQL/MariaDB, SQLite)
psycopg2-binary==2.9.10      # PostgreSQL driver
PyMySQL==1.1.1               # MySQL/MariaDB driver (pure Python)
# SQLite driver is built into Python stdlib - no package needed

# Cache
redis==5.2.1

# Web Server & WSGI
gunicorn==23.0.0
uvicorn[standard]==0.33.0

# Security & Authentication
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==44.0.1

# HTTP & API
requests==2.32.4
httpx==0.28.1
aiohttp==3.13.3

# Configuration
python-decouple==3.8
click==8.1.8

# Monitoring & Metrics
prometheus-client==0.21.1

# Logging
structlog==24.5.0

# Validation
jsonschema==4.23.0
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Development Dependencies (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.0
bandit==1.8.3
safety==3.3.1

# License Server Integration
# (Using requests for HTTP client - already included above)